Clarified names and comments regarding database history loading.
Renamed _load_database() to _load_history() in HarosAnalyseRunner
to better reflect what it is doing. Also added comments to give
some context and moved the setting of current_dir out of the
function, because its not specific to this function (used elsewhere
as well).
Also moved the comment about HarosDatabase._compact() to where
the database is saved, because that is where the _compact() is
performed.