Added new --minimal-output parameter to limit report size.
The current report output includes several files which are not
actually required for viewing the dashboard.
In order to avoid overhead when HAROS is used by many
packages on the ROS buildfarm and wiki, a new parameter
--minimal-output was added that, when set, supresses
the creation of unneccessary files. Currently, only the
creation of the haros.db is suppressed.