Added the external library js files back in, removal now done dynamically.
In order to allow offline-reports, the 3rd party js libraries were
added back into the harosviz/ folder, and instead a mechanism was
added to remove known libraries dynamically during the harosviz
installation process.
Also added documentation on how to reduce the font size so that
in the future new icons can be added easily.